Changed: Always wait 5 sec after sending save req before saving file settings [Stulle] - Stulle Changed: Write FileSettings.ini sequentially (faster) [Stulle] - Stulle Changed: Global SysInfo RAM stats will support usage of more than 4GB [Stulle] - Stulle Changed: Min small files push size to 1 kB [Stulle] - Stulle Changed: Small files push size slider kB based (makes keyboard input usable) [Stulle] - Stulle Fixed: Save known or FileSettings threads were not terminated properly [Stulle] - Stulle Fixed: Changing the small files push size reset changes in the Tree options [Stulle] - Stulle

+++ srchybrid/MiniMule/SystemInfo.cpp
@@ -23,6 +23,6 @@ int CSystemInfo::GetGlobalCpuUsage(){
 	return sys;
 }
 
-uint32 CSystemInfo::GetGlobalMemoryUsage(){
-	return (uint32)(CSysInfo::GetTotalPhys()-CSysInfo::GetAvailPhys());
+uint64 CSystemInfo::GetGlobalMemoryUsage(){
+	return CSysInfo::GetTotalPhys()-CSysInfo::GetAvailPhys();
 }
